feat: use DRSM only when database is enabled (#270)
* feat: use DRSM only when database is enabled


* fix missing clauses


* improve handling of the casts


* initialize local generators only when database is disabled


* Revert "improve handling of the casts"

This reverts commit a2d5fe5a402b2f8c2874fdc40610a47201c5f163.


* address review comments


---------

+++ oam/api_purge_ue_context.go
@@ -57,7 +57,9 @@ func HTTPAmfInstanceDown(c *gin.Context) {
 	if nfInstanceId, exists := c.Params.Get("nfid"); exists {
 		req.Params["nfid"] = nfInstanceId
 		self := context.AMF_Self()
-		self.Drsm.DeletePod(nfInstanceId)
+		if self.EnableDbStore {
+			self.Drsm.DeletePod(nfInstanceId)
+		}
 		c.JSON(http.StatusOK, nil)
 	}
 }
